pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "python-telegram-bot==13.7",
    "sqlalchemy>=2.0.38",
    "telegram>=0.0.1",
]
